+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - LVGL experimental mirroring of display on Web UI (#23041)
 - Allow acl in mqtt when client certificate is in use with `#define USE_MQTT_CLIENT_CERT` (#22998)
 - Berry `tasmota.when_network_up()` and simplified Matter using it
+- Berry `introspect.isconst()` to know if a Berry object is solidified or in RAM
 
 ### Breaking Changed